+++ packages/db/src/schema/account.ts
@@ -1,0 +1,37 @@
+import { relations } from "drizzle-orm";
+import { pgTable, primaryKey } from "drizzle-orm/pg-core";
+
+import { User } from "./user";
+
+export const Account = pgTable(
+  "account",
+  (t) => ({
+    userId: t
+      .uuid()
+      .notNull()
+      .references(() => User.id, { onDelete: "cascade" }),
+    type: t
+      .varchar({ length: 255 })
+      .$type<"email" | "oauth" | "oidc" | "webauthn">()
+      .notNull(),
+    provider: t.varchar({ length: 255 }).notNull(),
+    providerAccountId: t.varchar({ length: 255 }).notNull(),
+    refresh_token: t.varchar({ length: 255 }),
+    access_token: t.text(),
+    expires_at: t.integer(),
+    token_type: t.varchar({ length: 255 }),
+    scope: t.varchar({ length: 255 }),
+    id_token: t.text(),
+    session_state: t.varchar({ length: 255 }),
+  }),
+  (account) => [
+    primaryKey({
+      name: "pk_account",
+      columns: [account.provider, account.providerAccountId],
+    }),
+  ],
+);
+
+export const AccountRelations = relations(Account, ({ one }) => ({
+  user: one(User, { fields: [Account.userId], references: [User.id] }),
+}));
