+++ packages/db/src/schema/session.ts
@@ -1,0 +1,17 @@
+import { relations } from "drizzle-orm";
+import { pgTable } from "drizzle-orm/pg-core";
+
+import { User } from "./user";
+
+export const Session = pgTable("session", (t) => ({
+  sessionToken: t.varchar({ length: 255 }).notNull().primaryKey(),
+  userId: t
+    .uuid()
+    .notNull()
+    .references(() => User.id, { onDelete: "cascade" }),
+  expires: t.timestamp({ mode: "date", withTimezone: true }).notNull(),
+}));
+
+export const SessionRelations = relations(Session, ({ one }) => ({
+  user: one(User, { fields: [Session.userId], references: [User.id] }),
+}));
